requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
python-socketio==5.11.0
eventlet==0.33.3
simple-websocket==1.0.0
windows-curses; platform_system == "Windows"